// pop-client/src/app/shared/openapi/model/tumor-marker-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TumorMarkerTumorProportionScoreChoices } from './tumor-marker-tumor-proportion-score-choices';
import { CodedConcept } from './coded-concept';
import { TumorMarkerNuclearExpressionStatusChoices } from './tumor-marker-nuclear-expression-status-choices';
import { Measure } from './measure';
import { TumorMarkerImmunohistochemicalScoreChoices } from './tumor-marker-immunohistochemical-score-choices';
import { TumorMarkerPresenceChoices } from './tumor-marker-presence-choices';
import { TumorMarkerImmuneCellScoreChoices } from './tumor-marker-immune-cell-score-choices';


export interface TumorMarkerCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient related to the tumor marker result
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the tumor marker was analyzed.
     */
    date: string;
    /**
     * The chemical or biological substance/agent that is analyzed.
     */
    analyte: CodedConcept;
    massConcentration?: Measure | null;
    arbitraryConcentration?: Measure | null;
    substanceConcentration?: Measure | null;
    fraction?: Measure | null;
    multipleOfMedian?: Measure | null;
    tumorProportionScore?: TumorMarkerTumorProportionScoreChoices | null;
    immuneCellScore?: TumorMarkerImmuneCellScoreChoices | null;
    combinedPositiveScore?: Measure | null;
    immunohistochemicalScore?: TumorMarkerImmunohistochemicalScoreChoices | null;
    presence?: TumorMarkerPresenceChoices | null;
    nuclearExpressionStatus?: TumorMarkerNuclearExpressionStatusChoices | null;
    /**
     * References to the neoplastic entities that are related or the focus of the tumor marker analysis.
     */
    relatedEntitiesIds?: Array<string>;
}
export namespace TumorMarkerCreate {
}