// pop-client/src/app/shared/openapi/model/tumor-marker-immune-cell-score-choices.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TumorMarkerImmuneCellScoreChoices = 'IC0' | 'IC1' | 'IC2' | 'IC3';

export const TumorMarkerImmuneCellScoreChoices = {

    Ic0: 'IC0' as TumorMarkerImmuneCellScoreChoices,

    Ic1: 'IC1' as TumorMarkerImmuneCellScoreChoices,

    Ic2: 'IC2' as TumorMarkerImmuneCellScoreChoices,

    Ic3: 'IC3' as TumorMarkerImmuneCellScoreChoices
};